// packages/contracts/typechain-types/DummyContract.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import {
  ethers,
  EventFilter,
  Signer,
  BigNumber,
  BigNumberish,
  PopulatedTransaction,
  BaseContract,
  ContractTransaction,
  CallOverrides,
} from "ethers";
import { BytesLike } from "@ethersproject/bytes";
import { Listener, Provider } from "@ethersproject/providers";
import { FunctionFragment, EventFragment, Result } from "@ethersproject/abi";
import type {
  TypedEventFilter,
  TypedEvent,
  TypedListener,
  OnEvent,
} from "./common";

export interface DummyContractInterface extends ethers.utils.Interface {
  functions: {
    "dummy()": FunctionFragment;
  };

  encodeFunctionData(functionFragment: "dummy", values?: undefined): string;

  decodeFunctionResult(functionFragment: "dummy", data: BytesLike): Result;

  events: {};
}

export interface DummyContract extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: DummyContractInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    dummy(overrides?: CallOverrides): Promise<[void]>;
  };

  dummy(overrides?: CallOverrides): Promise<void>;

  callStatic: {
    dummy(overrides?: CallOverrides): Promise<void>;
  };

  filters: {};

  estimateGas: {
    dummy(overrides?: CallOverrides): Promise<BigNumber>;
  };

  populateTransaction: {
    dummy(overrides?: CallOverrides): Promise<PopulatedTransaction>;
  };
}